requirements.txt
Django~=4.1.10
djangorestframework~=3.14.0
psycopg2==2.9.6
django_rq
pytest-django